backend/requirements.txt
fastapi>=0.115.9
uvicorn==0.23.2
pydantic>=2.7.4
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
uuid==1.30
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
groq>=0.12.0
